int/ast/export: moves out some functions to separate package.
