+++ internal/ast/export/export_test.go
@@ -1,4 +1,4 @@
-package astwutl
+package export
 
 import (
 	"fmt"
