+++ internal/evolution/genetics/mutation/v1/lines/remove_test.go
@@ -4,9 +4,9 @@ import (
 	"fmt"
 	"testing"
 
-	"tde/internal/ast/astwutl"
 	"tde/internal/ast/clone"
 	"tde/internal/ast/compare"
+	"tde/internal/ast/export"
 	"tde/internal/evolution/genetics/mutation/v1/models"
 
 	"github.com/kylelemons/godebug/diff"
@@ -26,11 +26,11 @@ func Test_RemoveLine(t *testing.T) {
 		t.Fatal(fmt.Errorf("act: %w", err))
 	}
 
-	codeForOriginal, err := astwutl.String(originalFuncDecl)
+	codeForOriginal, err := export.String(originalFuncDecl)
 	if err != nil {
 		t.Error("validation prep")
 	}
-	codeForModified, err := astwutl.String(modifiedFuncDecl)
+	codeForModified, err := export.String(modifiedFuncDecl)
 	if err != nil {
 		t.Error("validation prep")
 	}
