+++ internal/evolution/genetics/mutation/v1/lines/switch_test.go
@@ -5,9 +5,9 @@ import (
 	"go/ast"
 	"testing"
 
-	"tde/internal/ast/astwutl"
 	"tde/internal/ast/clone"
 	"tde/internal/ast/compare"
+	"tde/internal/ast/export"
 	"tde/internal/ast/find"
 	"tde/internal/ast/parse"
 	"tde/internal/evolution/genetics/mutation/v1/models"
@@ -43,11 +43,11 @@ func Test_SiblingSwap(t *testing.T) {
 		t.Fatal(fmt.Errorf("act: %w", err))
 	}
 
-	codeForOriginal, err := astwutl.String(originalFuncDecl)
+	codeForOriginal, err := export.String(originalFuncDecl)
 	if err != nil {
 		t.Error("validation prep")
 	}
-	codeForModified, err := astwutl.String(modifiedFuncDecl)
+	codeForModified, err := export.String(modifiedFuncDecl)
 	if err != nil {
 		t.Error("validation prep")
 	}
